requirements.txt
altair==4.2.0
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
backports.zoneinfo==0.2.1
binance-connector==1.17.0
blinker==1.5
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
constantly==15.1.0
cryptography==38.0.1
decorator==5.1.1
entrypoints==0.4
future==0.18.2
gevent==21.12.0
gitdb==4.0.9
GitPython==3.1.27
greenlet==1.1.3
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
incremental==21.3.0
Jinja2==3.1.2
jsonschema==4.16.0
MarkupSafe==2.1.1
msgpack==1.0.4
numpy==1.23.3
packaging==21.3
pandas==1.4.4
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
plotly==5.10.0
protobuf==3.20.2
psycopg2==2.9.3
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.8.0b3
Pygments==2.13.0
pymongo==4.2.0
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==23.2.1
requests==2.28.1
rich==12.5.1
semver==2.13.0
service-identity==21.1.0
six==1.16.0
smmap==5.0.0
streamlit==1.12.2
tenacity==8.0.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
Twisted==22.8.0
#twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wincertstore==0.2
zerorpc==0.6.3
zipp==3.8.1
zope.event==4.5.0
zope.interface==5.4.0
